Prevent model failure if level count is less than 34
The number of levels which can contain clouds (LWEPAR) is currently
hard-coded. Since GEOS-Chem can be run with different vertical
resolutions and driving models, this can cause an error. LWEPAR
is now set at runtime, and (as a temporary fix) is set to be the
minimum of the standard value (34) and the model level count. In
future a more robust solution will be needed.